move ai from background to jobfetcher
